requirements.txt
praw==7.2.0
prettytable==2.1.0
colorama==0.4.4
python-dotenv==0.17.1